refactor: source actions
