--- packages/shared/src/components/sources/SourceActions/SourceActionsNotify.tsx
@@ -1,35 +1,0 @@
-import React, { ReactElement } from 'react';
-import { ButtonSize, ButtonVariant } from '../../buttons/common';
-import { SimpleTooltip } from '../../tooltips/SimpleTooltip';
-import { Button } from '../../buttons/Button';
-import { BellAddIcon, BellSubscribedIcon } from '../../icons';
-
-interface SourceActionsNotifyProps {
-  haveNotifications: boolean;
-  onClick: () => void;
-}
-
-const SourceActionsNotify = (props: SourceActionsNotifyProps): ReactElement => {
-  const { haveNotifications, onClick } = props;
-
-  const icon = haveNotifications ? <BellSubscribedIcon /> : <BellAddIcon />;
-  const label = `${haveNotifications ? 'Disable' : 'Enable'} notifications`;
-  const variant = haveNotifications
-    ? ButtonVariant.Tertiary
-    : ButtonVariant.Secondary;
-
-  return (
-    <SimpleTooltip content={label}>
-      <Button
-        aria-label={label}
-        icon={icon}
-        onClick={onClick}
-        size={ButtonSize.Small}
-        title={label}
-        variant={variant}
-      />
-    </SimpleTooltip>
-  );
-};
-
-export default SourceActionsNotify;
