+++ packages/shared/src/components/sources/SourceActions/index.tsx
@@ -1,15 +1,22 @@
-import React from 'react';
-import {
-  SourceActionsWithNotifyExperiment,
-  SourceActionsProps,
-} from './SourceActionsWithNotifyExperiment';
-import { withExperiment } from '../../withExperiment';
-import { feature } from '../../../lib/featureManagement';
-import { SourceSubscribeButton } from '../SourceSubscribeButton';
+import React, { ReactElement } from 'react';
+import { SourceSubscribeButton } from './SourceSubscribeButton';
 import SourceActionsBlock from './SourceActionsBlock';
 import { useSourceActionsBlock } from '../../../hooks/source/useSourceActionsBlock';
+import { Source } from '../../../graphql/sources';
+import { ButtonVariant } from '../../buttons/common';
 
-const SourceDefaultActions = (props: SourceActionsProps) => {
+interface SourceActionsButton {
+  className?: string;
+  variant?: ButtonVariant;
+}
+
+interface SourceActionsProps {
+  source: Source;
+  hideBlock?: boolean;
+  followProps?: SourceActionsButton;
+}
+
+export const SourceActions = (props: SourceActionsProps): ReactElement => {
   const { source, hideBlock, followProps } = props;
   const { isBlocked, toggleBlock } = useSourceActionsBlock({
     source,
@@ -24,11 +31,5 @@ const SourceDefaultActions = (props: SourceActionsProps) => {
     </>
   );
 };
-
-export const SourceActions = withExperiment(SourceActionsWithNotifyExperiment, {
-  fallback: SourceDefaultActions,
-  feature: feature.sourceNotifyButton,
-  value: true,
-});
 
 export default SourceActions;
