+++ packages/shared/src/hooks/source/index.ts
@@ -1,2 +1,1 @@
-export * from './useSourceActions';
 export * from './useSourceActionsNotify';
